requirements.txt
streamlit==1.31.0
plotly==5.18.0
pandas==2.2.0
numpy==1.26.3
websocket-client==1.7.0
sqlalchemy==2.0.25
statsmodels==0.14.1
scikit-learn==1.4.0
scipy==1.12.0
redis==5.0.1
python-dotenv==1.0.0
loguru==0.7.2